module: add zephyr module dir and yaml
Temporarily add a basic `zephyr/module.yaml` entry to enable
integration with Zephyr's build system.

Zephyr module integration will be done in the main Zephyr tree.

This change itself should eventually go upstream and the
preference for this project would be to have all changes go
upstream first.

So in a sense, the fork of `apache/thrift` to
`zephyrproject-rtos/thrift` should be kept as redundant as
possible so that it could be replaced with the upstream project
at any time.

Signed-off-by: Chris Friedt <cfriedt@meta.com>